LayoutTests/media/modern-media-controls/media-controls/media-controls-display-above-captions.html

<script src="../../../resources/js-test-pre.js"></script>
<body>
<video src="../../content/CC+Subtitles.mov" controls></video>
<script type="text/javascript">

description("Testing that the media controls container is above the captions container.");

const media = document.querySelector("video");
const shadowRoot = window.internals.shadowRoot(media); 
const container = shadowRoot.querySelector(".media-controls-container");
shouldBe("container.childElementCount", "2");
shouldBeEqualToString("container.firstElementChild.getAttribute('useragentpart')", "-webkit-media-text-track-container");
shouldBeTrue("container.lastElementChild.classList.contains('media-controls')");

</script>
<script src="../../../resources/js-test-post.js"></script>
</body>
