LayoutTests/media/modern-media-controls/media-controls/media-controls-display-above-captions-expected.txt

Testing that the media controls container is above the captions container.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS container.childElementCount is 2
PASS container.firstElementChild.getAttribute('useragentpart') is "-webkit-media-text-track-container"
PASS container.lastElementChild.classList.contains('media-controls') is true
PASS successfullyParsed is true

TEST COMPLETE
